Import Debian 06-CVE-2007-1673.patch, CVE-2007-1673
Multiple vendors ZOO file decompression infinite loop DoS, CVE-2007-1673

It's possible to make the ZOO implementation to enter in an infinite
loop condition.  The vulnerability lies in the algorithm used to locate
the files inside the archive.  Each file in a ZOO archive is identified
by a direntry structure.  Those structures are linked between themselves
with a 'next' pointer.  This pointer is in fact an offset from the
beginning of the file, representing the next direntry structure.  By
specifying an already processed file, it's possible to process more than
one time this same file.  The ZOO parser will then enter an infinite
loop condition.

http://archives.neohapsis.com/archives/bugtraq/2007-05/0046.html

Signed-off-by: Jean-Sebastien Guay-Leroux <jean-sebastienguay-leroux.com>
Signed-off-by: Joachim Nilsson <troglobit@gmail.com>