zoo is now fully in the public domain
Thanks to the efforts of Niklas Vainio of Debian, Rahul Dhesi has put
the entirety of the zoo code base in the public domain.  See the new
file copyright.eml for the complete conversation.

This patch drops the file Copyright, and all copyright headers in
relevant files with copyright previously assigned to Rahul Dhesi.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>