Follow-up to cd71ff8: properly fix 64-bit long in zoo_header
This is the proper fix to the 64-bit long problem in `struct zoo_header`
that was "fixed" in Debian 03-fix-manage-archive-under-AMD64.patch.

The on-disk .zoo format is fixed and handled by marshalling functions
that ensure long is 32-bit (four bytes) and int is 16-bit (two bytes).
This patch only ensures the proper data types are used internally in
zoo to prevent sign extension problems.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>